fix(TUP-20926): extract OutputSchemaParameter
